requirements.txt
PyYAML>=6.0
matplotlib>=3.8
networkx>=3.2
scipy>=1.10
hydra-core>=1.3.2
